+++ src/ChunkCache.cc
@@ -25,12 +25,12 @@ bool ChunkCache::open(const fs::path& path, bool read_only)
     return true;
 }
 
-bool ChunkCache::load_chunk(std::uint32_t chunk_id)
+bool ChunkCache::load_chunk(std::uint64_t chunk_id)
 {
     if (!m_handler.seek(chunk_id * capacity))
         return false;
 
-    std::uint32_t bytes_to_read = capacity;
+    std::uint64_t bytes_to_read = capacity;
     if (chunk_id == (m_total_chunks - 1) && m_handler.size() % capacity)
         bytes_to_read = m_handler.size() % capacity;
