+++ src/Utilities.h
@@ -60,6 +60,6 @@ const char* get_arg(int argc, const char* const* const argv, const std::string& 
 
 bool get_flag(int argc, const char* const* const argv, const std::string& flag);
 
-std::uint32_t str_to_int(const char* const str);
+std::uint64_t str_to_int(const char* const str);
 } // namespace Hexit
 #endif // UTILITIES_H
